test(canbus-events): Add round-trip parser unit tests for CANaerospace schema and fix redundant mock in `CanaerospaceFrameParserTest`

+++ canbus-events/src/main/java/io/mapsmessaging/canbus/canaerospace/parser/CanaerospaceFrameParser.java
@@ -51,7 +51,7 @@ public class CanaerospaceFrameParser {
 
       Double resolution = resolvedIdentifier.getResolution();
       if (resolution != null && raw instanceof Number number) {
-        engineeringValue = number.doubleValue() * resolution.doubleValue();
+        engineeringValue = number.doubleValue() * resolution;
       } else if (raw instanceof Float floatValue) {
         engineeringValue = floatValue.doubleValue();
       } else if (raw instanceof Double doubleValue) {
